requirements.txt
# Machine Learning y Data Science
numpy==2.3.4
pandas==2.3.3
scikit-learn
joblib==1.5.2
seaborn==0.13.2
matplotlib==3.10.7

# Gráficos y visual
cycler==0.12.1
contourpy==1.3.3
fonttools==4.60.1
kiwisolver==1.4.9
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5

# Fechas y zonas horarias
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0

# Excel y bases de datos
openpyxl==3.1.5
mysql-connector-python

# Streamlit y utilidades web
streamlit

# Manejo de variables de entorno
python-dotenv

